Add TestOutputHelper.QueueTestOutput integration.
